src/sample-service/requirements.txt
asyncio-nats-client